+++ Academia/app/src/main/java/com/example/ravenmargret/academia/Scale.java
@@ -7,7 +7,7 @@ public class Scale implements Serializable
     private static final long serialVersionUID = 8736847634070552888L;
     String mLetterGrade;
     String mNumberGrade;
-    int mGPAValue;
+    double mGPAValue;
 
     public Scale(String mLetterGrade, String mNumberGrade, int mGPAValue)
     {
@@ -18,12 +18,12 @@ public class Scale implements Serializable
 
     public String getmLetterGrade() {return mLetterGrade;}
     public String getmNumberGrade() {return mNumberGrade;}
-    public int getmGPAValue() {return mGPAValue;}
+    public double getmGPAValue() {return mGPAValue;}
 
 
     public void setmLetterGrade(String mLetterGrade) {this.mLetterGrade = mLetterGrade;}
     public void setmNumberGrade(String mNumberGrade) {this.mNumberGrade = mNumberGrade;}
-    public void setmGPAValue(int mGPAValue) {this.mGPAValue = mGPAValue;}
+    public void setmGPAValue(double mGPAValue) {this.mGPAValue = mGPAValue;}
 
     @Override
     public String toString()
